put gui windows in seperate scrips, fixed textbox order in welcome window, added code to move also exisiting seg to in progress window
