+++ src/data_centric_platform/client/settings.py
@@ -1,0 +1,3 @@
+def init():
+    global accepted_types
+    accepted_types = (".jpg", ".jpeg", ".png", ".tiff", ".tif")
